config/init: fix failure in 'narrow' terminals
The output of 'docker-compose ps' command depends on the current widht
of the terminal window. When the window is too narrow the lines remain
wrapped in the variable which prevents the instructions using 'grep'
from working as expected.

'docker ps' output to a variable preserves lines structer regardles of
the terminal window widht.